Idiot-proofing
By idiot i mean me, and not my lovely users whom i love very much. I think I got the app to not explode when offline, but does not generate card images, which is to be expected. This program relies on Scryfall for its data and I don't want to store 32k png's on everyone's computers lol